// src/app/modules/core/authorization/enum/permission.enum.ts
export enum Permission {
  RESERVATION_VIEW = 'RESERVATION_VIEW',
  RESERVATION_VIEW_USER = 'RESERVATION_VIEW_USER',
  RESERVATION_EDIT = 'RESERVATION_EDIT',
  RESERVATION_EDIT_USER = 'RESERVATION_EDIT_USER',
  ROOM_VIEW = 'ROOM_VIEW',
  ROOM_EDIT = 'ROOM_EDIT',
  APPLIANCE_VIEW = 'APPLIANCE_VIEW',
  APPLIANCE_EDIT = 'APPLIANCE_EDIT',
  SOFTWARE_VIEW = 'SOFTWARE_VIEW',
  SOFTWARE_EDIT = 'SOFTWARE_EDIT',
  USER_VIEW = 'USER_VIEW',
  USER_EDIT = 'USER_EDIT'
}